+++ Project2-Stream-Compaction/src/testing_helpers.hpp
@@ -10,7 +10,7 @@ template<typename T>
 int cmpArrays(unsigned long int n, T *a, T *b) {
     for (unsigned long int i = 0; i < n; i++) {
         if (a[i] != b[i]) {
-            printf("    a[%d] = %d, b[%d] = %d\n", i, a[i], i, b[i]);
+            printf("    a[%lld] = %lld, b[%lld] = %lld\n", i, a[i], i, b[i]);
             return 1;
         }
     }
@@ -30,7 +30,7 @@ void printCmpResult(unsigned long int n, T *a, T *b) {
 template<typename T>
 void printCmpLenResult(unsigned long int n, unsigned long int expN, T *a, T *b) {
     if (n != expN) {
-        printf("    expected %d elements, got %d\n", expN, n);
+        printf("    expected %lld elements, got %lld\n", expN, n);
     }
     printf("    %s \n",
             (n == -1 || n != expN) ? "FAIL COUNT" :
@@ -38,19 +38,19 @@ void printCmpLenResult(unsigned long int n, unsigned long int expN, T *a, T *b) 
 }
 
 
-void zeroArray(unsigned long int n, int *a) {
+void zeroArray(unsigned long int n, long long *a) {
     for (unsigned long int i = 0; i < n; i++) {
         a[i] = 0;
     }
 }
 
-void onesArray(unsigned long int n, int *a) {
+void onesArray(unsigned long int n, long long *a) {
 	for (unsigned long int i = 0; i < n; i++) {
 		a[i] = 1;
 	}
 }
 
-void genArray(unsigned long int n, int *a, int maxval) {
+void genArray(unsigned long int n, long long *a, int maxval) {
     srand(time(nullptr));
 
     for (unsigned long int i = 0; i < n; i++) {
@@ -58,7 +58,7 @@ void genArray(unsigned long int n, int *a, int maxval) {
     }
 }
 
-void printArray(unsigned long int n, int *a, bool abridged = false) {
+void printArray(unsigned long int n, long long *a, bool abridged = false) {
     printf("    [ ");
     for (unsigned long int i = 0; i < n; i++) {
         if (abridged && i + 2 == 15 && n > 16) {
