+++ Project2-Stream-Compaction/stream_compaction/efficient.h
@@ -6,14 +6,14 @@ namespace StreamCompaction {
     namespace Efficient {
         StreamCompaction::Common::PerformanceTimer& timer();
 
-        void scan(unsigned long long int n, int *odata, const int *idata);
+        void scan(unsigned long long int n, long long *odata, const long long *idata);
 
-		unsigned long long int compact(unsigned long long int n, int *odata, const int *idata);
+		unsigned long long int compact(unsigned long long int n, long long *odata, const long long *idata);
     }
 	namespace SharedMemory {
 		StreamCompaction::Common::PerformanceTimer& timer();
 
-		void scan(unsigned long long int n, int *odata, int *idata);
-		void dev_scan(unsigned long long int n, int *odata, int *idata);
+		void scan(unsigned long long int n, long long *odata, long long *idata);
+		void dev_scan(unsigned long long int n, long long *odata, long long *idata);
 	}
 }
