+++ Project2-Stream-Compaction/stream_compaction/thrust.cu
@@ -17,16 +17,16 @@ namespace StreamCompaction {
         /**
          * Performs prefix-sum (aka scan) on idata, storing the result into odata.
          */
-        void scan(unsigned long long int n, int *odata, const int *idata) {
+        void scan(unsigned long long int n, long long *odata, const long long *idata) {
             timer().startGpuTimer();
-			int *dev_odata, *dev_idata;
-			cudaMalloc((void**)&dev_odata, n * sizeof(int));
-			cudaMalloc((void**)&dev_idata, n * sizeof(int));
-			cudaMemcpy(dev_idata, idata, n * sizeof(int), cudaMemcpyHostToDevice);
-			thrust::device_ptr<int> dv_in(dev_idata);
-			thrust::device_ptr<int> dv_out(dev_odata);
+			long long *dev_odata, *dev_idata;
+			cudaMalloc((void**)&dev_odata, n * sizeof(long long));
+			cudaMalloc((void**)&dev_idata, n * sizeof(long long));
+			cudaMemcpy(dev_idata, idata, n * sizeof(long long), cudaMemcpyHostToDevice);
+			thrust::device_ptr<long long> dv_in(dev_idata);
+			thrust::device_ptr<long long> dv_out(dev_odata);
             thrust::exclusive_scan(dv_in, dv_in + n, dv_out);
-			cudaMemcpy(odata, dev_odata, sizeof(int) * n, cudaMemcpyDeviceToHost);
+			cudaMemcpy(odata, dev_odata, sizeof(long long) * n, cudaMemcpyDeviceToHost);
 			cudaFree(dev_odata);
 			cudaFree(dev_idata);
             timer().endGpuTimer();
