+++ Project2-Stream-Compaction/stream_compaction/thrust.h
@@ -6,6 +6,6 @@ namespace StreamCompaction {
     namespace Thrust {
         StreamCompaction::Common::PerformanceTimer& timer();
 
-        void scan(unsigned long long int n, int *odata, const int *idata);
+        void scan(unsigned long long int n, long long *odata, const long long *idata);
     }
 }
